Add different scenarios for bridging
There are at least two typical configurations for bridge mode.  UPSTREAM: The Optus is upstream from a router, and that router wants to get a DHCP address from Optus, and be responsible for WiFi. DOWNSTREAM: The Optus is providing WiFi and/or Ethernet fan-out but DHCP comes from the upstream main router.  This change explains how to do the DOWNSTREAM scenario.   I'm using this to replace Optus NBN with Starlink while retaining my router. 

Also confirmed the optus/optus non admin password, and admin/0ptU%1M5 but unable to check firmware version number as cannot access without a factory reset and starting again, it was 10.x